pyproject.toml
[project]
name = "business-assistant-project-management-plugin"
version = "0.1.0"
description = "Project management orchestration plugin for Business Assistant v2"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "business-assistant-v2",
    "sqlalchemy",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/business_assistant_pm"]

[tool.uv.sources]
business-assistant-v2 = { path = "D:/GIT/BenjaminKobjolke/business-assistant-v2", editable = true }

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.4.0",
    "mypy>=1.0.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "SIM"]

[tool.mypy]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
